docker-hadoop-spark-master/api/requirements.txt
fastapi
uvicorn
pyjwt
pyarrow
streamlit
pandas
requests